// core/src/shared/enums/tokens.ts
export enum TokenValues {
  LOCAL_ACCESS_TOKEN = "LOCAL_ACCESS_TOKEN",
  LOCAL_REFRESH_TOKEN = "LOCAL_REFRESH_TOKEN",
  LOCAL_USER_UUID = "LOCAL_USER_UUID",
  LOCAL_USER_USERNAME = "LOCAL_USER_USERNAME",
  LOCAL_USER_GUEST_STATUS = "LOCAL_USER_GUEST_STATUS",
  COMMUNITY_CONFIGURATIONS = "COMMUNITY_CONFIGURATIONS",
}