// core/src/shared/enums/wordAction.ts
export enum WordAction {
  FIRST_LETTER_CAPITAL_SINGULAR = "FIRST_LETTER_CAPITAL_SINGULAR",
  ALL_CAPITAL_SINGULAR = "ALL_CAPITAL_SINGULAR",
  ALL_SMALL_SINGULAR = "ALL_SMALL_SINGULAR",
  FIRST_LETTER_CAPITAL_PLURAL = "FIRST_LETTER_CAPITAL_PLURAL",
  ALL_CAPITAL_PLURAL = "ALL_CAPITAL_PLURAL",
  ALL_SMALL_PLURAL = "ALL_SMALL_PLURAL",
}